General refactors, changes to project structure
The packages in the /internal directory have been merged with the ones in the project root. Rewritten the slash commands handler to reduce complexity. The `CommandHandler` struct now serves as a shared state between all slash command handlers. All filter and event handling functionality has been moved to the main package. Changed config and database package to look for files relative to the executable instead of relative to the working directory. Also got rid of most if not all global variables, the config is now loaded in the main function and passed to other packages via there.